codebert-base.commons-lang.mid-2239.idx-18986.2.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -240,7 +240,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      * @since 3.5
      */
     public float addAndGet(final float operand) {
-        this.value += operand;
+        this.value += Double.valueOf(operand).floatValue();
         return value;
     }
